// src/Enums/enum.ts
export enum ItemType {
  AgedBrie = 'Aged Brie',
  BackstagePasses = 'Backstage passes to a TAFKAL80ETC concert',
  Sulfuras = 'Sulfuras, Hand of Ragnaros',
}
  
export enum Quality {
  MIN = 0,
  MAX = 50,
  INCREASE = 1,
  INCREASE_DOUBLE = 2,
  DECREASE = 1,
}
  
export enum SellIn {
  WARNING = 11,
  SOON = 6,
}